+++ R/mod_analysisSettings_PhenotypeScoring.R
@@ -1,0 +1,275 @@
+#' @title PhenotypeScorings Analysis Settings UI
+#' @description UI module for configuring the settings for cohort overlaps analysis. This module allows users to select cohorts and set the minimum cell count for the analysis.
+#'
+#' @param id A string representing the module's namespace.
+#'
+#' @return A Shiny UI element that can be included in a Shiny app.
+#'
+#' @importFrom shiny NS
+#' @importFrom shiny tags
+#' @importFrom shiny h4
+#' @importFrom shinyWidgets pickerInput
+#' @importFrom shinyjs useShinyjs
+#' @importFrom htmltools tagList
+#'
+#' @export
+#'
+mod_analysisSettings_phenotypeScoring_ui <- function(id) {
+  ns <- shiny::NS(id)
+  htmltools::tagList(
+    shinyjs::useShinyjs(),
+
+    shiny::fluidRow(
+      # Cohorts panel
+      shiny::column(
+        width = 6,
+        shiny::wellPanel(
+          shiny::tags$h4("Cohorts"),
+          shinyWidgets::pickerInput(
+            inputId = ns("selectCaseCohort_pickerInput"),
+            label = "Select case cohort:",
+            choices = NULL,
+            selected = NULL,
+            multiple = FALSE
+          ),
+          shinyWidgets::pickerInput(
+            inputId = ns("selectControlCohort_pickerInput"),
+            label = "Select control cohort:",
+            choices = NULL,
+            selected = NULL,
+            multiple = FALSE
+          )
+        )
+      ),
+
+      # Upload panel
+      shiny::column(
+        width = 6,
+        shiny::wellPanel(
+          shiny::tags$h4("Upload Previous Result"),
+          shiny::helpText(
+            "Upload a previously run codeWAS result file (.duckdb):"
+          ),
+          shiny::fileInput(
+            inputId = ns("upload_result_duckdb"),
+            label = NULL,
+            accept = c(".duckdb", ".db"),
+            buttonLabel = "Browse...",
+            placeholder = "No file selected"
+          ),
+          shiny::actionButton(
+            ns("viewPreviousResult_actionButton"),
+            "Open previous result for phenotype scoring",
+            class = "btn-primary"
+          )
+        )
+      )
+    ),
+
+    htmltools::hr(),
+
+    shiny::tags$h4("Settings"),
+    shiny::fluidRow(
+      # run codeWAS settings
+      shiny::column(
+        width = 12,
+        mod_fct_covariateSelector_ui(
+          inputId = ns("features_pickerInput"),
+          label = "Select features to compare between cases and controls:",
+          analysisIdsToShow = c(
+            1, 2, 10, 41,
+            101, 141,
+            301, 342,
+            501, 541,
+            701, 741,
+            801, 841,
+            601, 641
+          ),
+          analysisIdsSelected = c(1, 2, 10, 41, 141, 342)
+        )
+      )
+    ),
+
+    htmltools::hr(),
+
+    shiny::tags$h4("Pre-run info"),
+    shiny::verbatimTextOutput(ns("info_text"), placeholder = TRUE)
+  )
+}
+
+
+#' @title PhenotypeScorings Analysis Settings Server
+#' @description Server module for handling the logic of the cohort overlaps analysis settings UI. This module updates the UI elements based on the selected cohorts and returns the analysis settings as a reactive expression.
+#'
+#' @param id A string representing the module's namespace.
+#' @param r_connectionHandler A reactive object that provides access to the cohort data and connection information.
+#'
+#' @return A reactive expression that returns the analysis settings as a list.
+#'
+#' @importFrom shiny moduleServer
+#' @importFrom shiny reactive
+#' @importFrom shiny req
+#' @importFrom shiny observe
+#' @importFrom shinyjs toggleState
+#' @importFrom shinyWidgets updatePickerInput
+#'
+#' @export
+#'
+mod_analysisSettings_phenotypeScoring_server <- function(id, r_connectionHandler) {
+  shiny::moduleServer(id, function(input, output, session) {
+    ns <- session$ns
+
+    # TEMP
+    cores <- 1#shiny::getShinyOption("cores")
+    chunksSizeNOutcomes <- 5000#shiny::getShinyOption("chunksSizeNOutcomes")
+    # END TEMP
+
+    #
+    # reactive variables
+    #
+
+    #
+    # update selectCohorts_pickerInput with cohort names in r_connectionHandler
+    #
+    shiny::observe({
+      shiny::req(r_connectionHandler$cohortTableHandler)
+      shiny::req(r_connectionHandler$hasChangeCounter)
+
+      cohortIdAndNames <- r_connectionHandler$cohortTableHandler$getCohortIdAndNames()
+      cohortIdAndNamesList <- list()
+      if(nrow(cohortIdAndNames) != 0){
+        cohortIdAndNamesList <- as.list(setNames(cohortIdAndNames$cohortId, paste(cohortIdAndNames$shortName, "("  , cohortIdAndNames$cohortName, ")")))
+      }
+
+      shinyWidgets::updatePickerInput(
+        inputId = "selectCaseCohort_pickerInput",
+        choices = cohortIdAndNamesList,
+        selected = character(0)
+      )
+    })
+
+    #
+    # Enable previous result Viewer button only if a file has been uploaded
+    #
+
+    shiny::observe({
+      file_uploaded <- shiny::isTruthy(input$upload_result_duckdb)
+      shinyjs::toggleState("viewPreviousResult_actionButton", condition = file_uploaded)
+    })
+
+    #
+    # if file is uploaded open a tab of visualization
+    #
+    shiny::observeEvent(input$viewPreviousResult_actionButton, {
+      shiny::req(input$upload_result_duckdb)
+
+      tryCatch({
+        # Normalize and encode path
+        uploaded_file <- normalizePath(input$upload_result_duckdb$datapath, winslash = "/")
+        uploaded_file <- utils::URLencode(uploaded_file)
+
+        if (!file.exists(input$upload_result_duckdb$datapath)) {
+          stop("Uploaded file not found! It may have been deleted.")
+        }
+
+        # Construct URL (this has to come from the calling module, and should be sth like url_visualiseResults, hard coded for now )
+        base_url <- "http://localhost:8561"
+        url <- paste0(
+          base_url, "/?analysisType=phenotypeScoring&pathToResultsDatabase=",
+          uploaded_file
+        )
+
+
+        message("Attempting to open URL: ", url)
+        shinyjs::runjs(paste0("window.open('", url, "')"))
+
+      }, error = function(e) {
+        shiny::showNotification(paste("Error:", e$message), type = "error")
+      })
+    })
+
+    #
+    # update matchToCohortId_pickerInput with cohort names not in selectCaseCohort_pickerInput
+    #
+    shiny::observe({
+      shiny::req(r_connectionHandler$hasChangeCounter)
+      shiny::req(input$selectCaseCohort_pickerInput)
+
+      cohortIdAndNames <- r_connectionHandler$cohortTableHandler$getCohortIdAndNames()
+      cohortIdAndNamesList <- list()
+      if(nrow(cohortIdAndNames) != 0){
+        cohortIdAndNamesList <- as.list(setNames(cohortIdAndNames$cohortId, paste(cohortIdAndNames$shortName, "("  , cohortIdAndNames$cohortName, ")")))
+      }
+
+      cohortIdAndNamesList <- cohortIdAndNamesList |>
+        purrr::discard(~.x %in% input$selectCaseCohort_pickerInput)
+
+      # Add cohort 0 with
+      cohortIdAndNamesList <- c(list(`AUTO-MATCH: Creates a control cohort from the patiens not in case cohort that matches case cohort by sex and birth year with ratio 1:10` = 0), cohortIdAndNamesList)
+
+      shinyWidgets::updatePickerInput(
+        inputId = "selectControlCohort_pickerInput",
+        choices = cohortIdAndNamesList,
+        selected = 0
+      )
+    })
+
+    #
+    # activate settings if cohors have been selected
+    #
+    shiny::observe({
+      condition <- !is.null(input$selectControlCohort_pickerInput) & input$selectControlCohort_pickerInput!="NA"
+      #shinyjs::toggleState("selectCovariates", condition = condition )
+      shinyjs::toggleState("features_pickerInput", condition = condition )
+    })
+
+
+    #
+    # return reactive options
+    #
+    rf_analysisSettings <- shiny::reactive({
+      if(
+        is.null(input$selectCaseCohort_pickerInput) |
+        is.null(input$selectControlCohort_pickerInput) |
+        is.null(input$features_pickerInput)
+      ){
+        return(NULL)
+      }
+
+      analysisIds  <-  input$features_pickerInput |> as.numeric()
+
+      analysisSettings <- list(
+        cohortIdCases = input$selectCaseCohort_pickerInput |> as.integer(),
+        cohortIdControls = input$selectControlCohort_pickerInput |> as.integer(),
+        analysisIds = analysisIds
+      )
+
+      return(analysisSettings)
+
+    })
+
+    return(rf_analysisSettings)
+
+  })
+}
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+