+++ tests/testthat/test-mod_analysisSettings_CodeWAS.R
@@ -2,7 +2,7 @@
 test_that("mod_analysisSettings_CodeWAS works", {
 
   # set up
-  cohortTableHandler <- helper_createNewCohortTableHandler(addCohorts = "HadesExtrasFractureCohorts")
+  cohortTableHandler <- helper_createNewCohortTableHandler(addCohorts = "HadesExtrasFractureCohortsMatched")
   withr::defer({rm(cohortTableHandler);gc()})
 
   r_connectionHandler <- shiny::reactiveValues(
@@ -29,7 +29,7 @@ test_that("mod_analysisSettings_CodeWAS works", {
         minCellCount_numericInput = 1)
 
       analysisSettings <- rf_analysisSettings()
-    
+
       analysisSettings |> assertAnalysisSettings_CodeWAS() |> expect_no_error()
       analysisSettings |> expect_equal(
         list(
@@ -49,7 +49,8 @@ test_that("mod_analysisSettings_CodeWAS works", {
       )
 
       output$info_text |> expect_match("No subjects overlap between case and control cohorts")
-      output$info_text |> expect_match("There is a significant difference in year of birth distribution between case and control cohorts")
+      output$info_text |> expect_match("There is a significant difference in the shapes of year of birth distributions|There is a significant difference in the mean year of birth|There is significant difference both in the mean year of birth")
+
 
       # check "full" gets covariates
       session$setInputs(
@@ -81,9 +82,46 @@ test_that("mod_analysisSettings_CodeWAS works", {
         )
       )
 
+  
       output$info_text |> expect_match("No subjects overlap between case and control cohorts")
-      output$info_text |> expect_no_match("There is a significant difference in sex distribution between case and control cohorts")
-      output$info_text |> expect_no_match("There is a significant difference in year of birth distribution between case and control cohorts")
+
+
+      # "test subject-level comparison for matched cohorts"
+      session$setInputs(
+        selectCaseCohort_pickerInput = 1,
+        selectControlCohort_pickerInput = 2001,
+        features_pickerInput = c(101, 141, 1, 2, 402, 702, 41),
+        statistics_type_option =  "aggregated",
+        controlSex_checkboxInput = TRUE,
+        controlYearOfBirth_checkboxInput = TRUE,
+        minCellCount_numericInput = 1)
+
+      analysisSettings <- rf_analysisSettings()
+
+      analysisSettings |> assertAnalysisSettings_CodeWAS() |> expect_no_error()
+      analysisSettings |> expect_equal(
+        list(
+          cohortIdCases = 1,
+          cohortIdControls = 2001,
+          analysisIds = c(101, 141, 1, 2, 402, 702, 41),
+          covariatesIds = NULL,
+          minCellCount = 1,
+          chunksSizeNOutcomes = 5000,
+          cores = 1,
+          analysisRegexTibble = tibble::tribble(
+            ~analysisId, ~analysisName, ~analysisRegex,
+            999, "Endpoints", "^(?!.*\\[CohortLibrary\\]).*_case$",
+            998, "CohortLibrary", ".*\\[CohortLibrary\\]"
+          )
+        )
+      )
+
+    
+      output$info_text |> expect_match("No subjects overlap between case and control cohorts")
+      output$info_text |> expect_match("mean year of birth: Cases=.*Controls=.*")
+      output$info_text |> expect_match("There is a significant difference in the shapes of year of birth distributions|There is a significant difference in the mean year of birth|There is significant difference both in the mean year of birth")
+
+
 
     }
   )